+++ lib/src/model.dart
@@ -170,54 +170,6 @@ class EasyTableModel<ROW> extends ChangeNotifier {
     return w;
   }
 
-  double get pinnedColumnsWidth {
-    double w = 0;
-    for (EasyTableColumn column in _columns) {
-      if (column.pinned) {
-        w += column.width;
-      }
-    }
-    return w;
-  }
-
-  double get unpinnedColumnsWidth {
-    double w = 0;
-    for (EasyTableColumn column in _columns) {
-      if (!column.pinned) {
-        w += column.width;
-      }
-    }
-    return w;
-  }
-
-  double get allColumnsWidth {
-    double w = 0;
-    for (EasyTableColumn column in _columns) {
-      w += column.width;
-    }
-    return w;
-  }
-
-  int get unpinnedColumnsLength {
-    int v = 0;
-    for (EasyTableColumn column in _columns) {
-      if (!column.pinned) {
-        v++;
-      }
-    }
-    return v;
-  }
-
-  int get pinnedColumnsLength {
-    int v = 0;
-    for (EasyTableColumn column in _columns) {
-      if (column.pinned) {
-        v++;
-      }
-    }
-    return v;
-  }
-
   /// Revert to original sort order
   void clearSort() {
     _sortedColumns.clear();